Nav fixes so that the javascript is more general, and not based on IDs. Thus, you could have multiple navs on the same page. Still some work to do
