+++ index.html
@@ -41,7 +41,6 @@
 		</div>
 	</nav>
 
-
 	<script src="./dist/main.bundle.js"></script>
 </body>
 </html>
