+++ public/scss/partials/_navigation.scss
@@ -85,7 +85,7 @@
 }
 
 .navbar__menu-toggle{
-	display: block;
+	display: none;
 
 	span {
 		display: block;
